+++ d3x-morpheus-core/src/main/java/com/d3x/morpheus/array/mapped/MappedArrayConstructor.java
@@ -72,7 +72,7 @@ public class MappedArrayConstructor implements ArrayFactory.Constructor {
 
 
     @Override
-    public <T> Array<T> apply(Class<T> type, int length, double fillPct, T defaultValue) {
+    public <T> Array<T> apply(Class<T> type, int length, float fillPct, T defaultValue) {
         return apply(type, length, defaultValue, null);
     }
